restructure the hint dump to be more robust
Credits to jaq for the regex string used in line 222

Use the var sign to test for the sign names for Agitha and Jovani to be more reliable

Restructure Jovani to be consistent, even when the input may not be.

(Agitha is broken rn because she doesn't account for blank sign #16 )

Normal not-foolish-hints assume that they always start with 'They say that', and also aren't yet used. [Soon*TM*]