attack the dump_spoiler_log function
It was terribly written, it got on my nerves, it's not doing its function and only its function.
Given I may be using the spoiler log for other purposes soon, and the recent issue, I wanted to go ahead and tackle this. (also I am ill, lemme get low-hanging fruit)

(And now i pass #59 on to the new function! WOO!)

resolves #53