Pull the correct branch base over (#92)
* reset up my dev branch

* fixes #63

Woo, simple fix i overlooked.

* please actually delete the tabs when they need to be

* implement #75

Told you it would be easy

* fix #73

Apparently that line got removed. GR.

* fix #74

*grab all the low hanging fruit!*

* fix #68

* remove a double defintion

Why was that here?

* save my test suite

* fix #70 and #71

(And also write these functions better! Woo!)

* Revert "save my test suite"

This reverts commit 993c582c87d8c42e641d512342ceca40e70da05b.

* remove findall_to_list

It is no longer used

* update an issue on Jovani's to-do list

* Delete data.txt

How did you survive this long? LOL

* remember to remove {} from the rewards (oh well)

* clears the notebook upon reset

The janky imports are because of the janky way that I am resetting the tracker causing janky circular imports unintentionally

#77

* remove conflicting files for #83

I'm probably going to have to delete even more

* Renaming Files (#83)

* remove Hints.py so i can rename it

This is going to be a massive pull request

* Delete Hints.py

* rename Hints.py to hints.py

(And update the calls)

* rename SpoilerLog.py to spoiler_log.py

(and correct all calls to it)

* remove globals to rename it

* rename Globals.py to globals.py

(and update all calls)

* remove Utils to rename it

* rename data/Utils.py to data/utils.py

(and fix all calls to it)

* remove agitha for renaming

* rename Agitha.py to agitha.py

(and update all calls to it)

* remove Jovani for renaming

* rename Jovani.py to jovani.py

(and update all calls to it)
(and remove an old import that no longer is used)

* remove parent for renaming

* rename Parent.py to parent.py

(and update all calls to it)

* rename MainPage.py to main_page.py

(and update all calls to it)

* rename PickSpoiler.py to pick_spoiler.py

(and update all calls to it)

* rename ResetTracker.py to reset_tracker.py

(and update all calls to it)

* remove gui/Utils.py for renaming

* rename gui/Utils.py to gui/utils.py

(and update all calls to it)

* rename hint tool.py to hint_tool.py

* remove outdated TODOs

* clean up hint_tool.py

In preparation for the big refactoring

* Pep8 compliance (1) (#85)

* note why these are not pep8 compliant but will not be updated

* Update globals.py

yay! Simple issue.

* Update utils.py

yay! simple fix!

* Update hints.py

yay! Simple fix!

* Update spoiler_log.py

yay! simple fix!

* Update main_page.py

https://stackoverflow.com/a/37489941

Such, an annoying, PEP8, rule.

This *will* be fixed by #78 or so help me, I'm going to scream, I refuse to do a class for this! Yes a class would make this better BUT I DO NOT CARE!
OOP IS NOT THE ONE TRUE WAY!

* Update pick_spoiler.py

Another case of annoyed with PEP8

* Update reset_tracker.py

back to easy fixes, yay...

* Update utils.py

more easy fixes

* Update agitha.py

easy fix

* Update jovani.py

easy fixes

* Update parent.py

all easy fixes.. yay...

* save a note for future reference

Going to