create create_notepad, update_data_tabs, and reset_tab (read comment)
... Because reasons.

Creating the data tabs for the first time will be the only time I ever create the tab itself. But when I reset the tabs to their default states, I will be creating a default notepad again. (function 1).
Whenever I change the contents of a tab, I will be updating the data tabs, but also resetting the tab (by destroying what is already within.)