Release 0.2 (#100)
* completely wipe the branch

I want to start from scratch.

* create a rough list of what the app is doing currently

No optimizations, just explain what it's doing

* set up everything to get the main notebook

Easy step, sure. But it's a step.

* create the default notepads

So, so, simple. I hope.

* create create_notepad, update_data_tabs, and reset_tab (read comment)

... Because reasons.

Creating the data tabs for the first time will be the only time I ever create the tab itself. But when I reset the tabs to their default states, I will be creating a default notepad again. (function 1).
Whenever I change the contents of a tab, I will be updating the data tabs, but also resetting the tab (by destroying what is already within.)

* create functions for creating and closing tabs

Because I'm an idiot and YES tabs will be closed- *BY RACE MODE!*

Small issue of "if the tab doesn't exist, then closing it won't work (it'll raise an error)", and "if the tab exists, making it won't work (it'll raise an error)", so also a baby function to test if tab even exists.

* finally implement the reset function

Yay! Now I have enough pieces to move forward! FINALLY!

* add close_all_tabs, then DRY out the code some

(I have my reasons)

* add in the race mode button

Also:
- if add tabs is trying to use the try-except, it breaks and just doesn't create a tab.
- But if close tabs is using the data dict, it doesn't work, and doesn't close the tabs.

Spaghet.

oh also woops it forces options at the front. Huh. didn't think about that.

* forces option to always be last

(Easy fix!)

* get resetting working

So yeah now I am closing the tab entirely. Oh well.

* begin fiddling around with the spoiler log frame

#65 is being worked on here. I really hate this.

* rename the tabs for minimum size reasons

(Also because some people don't know who agitha or jovani are)

Thx to ninjano for the suggestion.

* uncouple options and spoiler log

Spoiler log can be its own tab now! WOO!

Adjusted minimum window size, in part because I felt that it was far too small, in part to accommodate this new tab.

* get the basic dropdown working

(Not fully functional, but it DOES work. Next up is getting the parsing working)

* begin work on the parsing.. by changing the root window title name.

.. Listen, it's a nice to have feature. xD

* shove everything for the parser into a class

Hey look, I'm doing an OOP thing! :D

(I also updated the to-do list, and made some progress towards the actual parsing)

* more correctly make the .json file again

there ya go, jaq, hopefully this is more robust.

* add the basic parsing

It's finally time. It's finally, time, for the big two.

I'm going to go ahead and push this forward because I foresee them being problems.

#65 Is complete.

* update README

I realized it was woefully outdated, and I pushed a quick fix.

* provide an error if the user forgets to input a log

woops, i got too excited for agitha and jovani!

* make a more clean warning page for no logs.

The buttons being highlighted bugs me but that's a UI update.

* add missing newlines

Sorry @jaqque *hehe*

* make the spoiler log picker update every call

-.- woops, minor bug

* fix the permission denied response

(it wasn't doing nothing because it wasn't set up correctly. My bad.)

* get the basic agitha handling in

*But not item collection, yet*
This is written super broken down because I was reading the pseudocode as I went, and noticed things that would be replicated- i have a parent class in mind.

* add in the basic item collection

That was.. yeah, i stole it from the old code; it works perfectly fine, needs no change.

* create the parent class for Agitha and Jovani

(and now we see why the last patch was so so broken up)

* readjust to accommodate Jovani being a bit strange

Jovani creates a checklist *then* disables bad ones; this is a bit weird and breaks things a bit.

* fully flesh out Jovani's tab

There may be lurking bugs- oh, I expect lurking bugs- but she's alive! She's fully functional at the default state!

There's a weird bug where Jovani inherits agitha's checkboxes, idk why, but it was a thing.

The code could probably definitely be cleaner. But Jovani does feel more readable.

* lazily fix the completion text

It doesn't need to be anything fancy

* forces the notes tab to be default after each reset

* update the checklist to actually use a scrolled frame

(Y'know, the whole point of this?)

* adjust where the reset tracker was for the spoiler log

it might be better to do it immediately, and move the set-notes-to-priority into its own function for readability, then put that here; for now, this works.

* have the notepad return the user to default notes on any reset

And smartly!

* Delete pseudo.py

Not needed anymore

* properly store the new tab's contents

Right, I have this dict for a reason.

* fix the master frame to actually fit the window

GR. Minor fixes

* add the requirements necessary for this program

Thank you @jaqque for pointing this out to me.

* Update README.md

Also thank you to @jaqque for aiding the update.

* remove a debug file from the soon release

(I had 2 test pys this week because test1 held a bunch of stuff for the old shopping tabs. this was supposed to be ignored and/or removed.