fix #104
1. this is poorly written
2. It was caused by me forgetting I need to grab the bool now.

I should just have the warning as its own function *THEN* the rest of stuff because it seems strange to just have it like this (hi am tired can not do words)

YES this would be resolved by the tab flickering issues, BUT. ... I wanted to have it done in case I do not get around to them by tomorrow. This was the biggest issue.