abstract out the separate chunks of saving (pt 1)
The length of the file was really getting on my nerves, especially when most of it was gathering the data and the other chunk was writing.

This one focuses solely on the getting the data out. While the og file was renamed and then ripped apart, it is not hooked up and working, so yeah, this code doesn't work- honestly, do not care.

- Updated return_child_widgets typing so the formatting in VSCode worked.
- Pulled grab_data out of a function into the main loop.
- renamed "save_notepad" and "save_checklist" to instead be "parse" instead of "save"
- renamed "return_first_child" to be "widget" instead of "child"